test: densify naval panel part suites (Slice E #4183)
Consolidate combine/locate/split helpers into naval_units_panel_test_support
and scenario factories into naval_units_panel_test_scenarios; re-measure
stay-split family (~1487 combined physical lines, ~33% below baseline).

Refs #4183
