+++ app/test/naval_units_panel_part3_test.dart
@@ -1,6 +1,5 @@
 // Tests for NavalUnitsPanel. SPEC/ui/naval-units-panel.md.
 
-import 'package:colonizethis_data/colonizethis_data.dart' show MapTopology;
 import 'package:colonizethis_logic/colonizethis_logic.dart' show homeFleetIdFor;
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart' show suppressLogsForTests;
@@ -14,165 +13,6 @@ import 'package:colonizethis_app/widgets/ct_transfer_list.dart';
 
 import 'naval_units_panel_test_support.dart';
 import 'widget_test_assets.dart';
-
-const String _mergePort = 'oldWorld|mergeport';
-
-Future<void> _tapFleetCheckboxes(
-  WidgetTester tester,
-  Iterable<String> fleetLabels, {
-  bool scroll = false,
-}) async {
-  for (final label in fleetLabels) {
-    late final Finder cb;
-    if (scroll) {
-      final title = find.text(label);
-      await tester.scrollUntilVisible(title, 120);
-      await tester.pumpAndSettle();
-      final tile = find.ancestor(
-        of: title,
-        matching: find.byType(ExpansionTile),
-      );
-      cb = find.descendant(of: tile, matching: find.byType(Checkbox));
-      await tester.scrollUntilVisible(cb, 120);
-      await tester.pumpAndSettle();
-    } else {
-      final tile = find.widgetWithText(ExpansionTile, label);
-      expect(tile, findsOneWidget);
-      cb = find.descendant(of: tile, matching: find.byType(Checkbox));
-    }
-    await tester.ensureVisible(cb);
-    await tester.tap(cb);
-    await tester.pumpAndSettle();
-  }
-}
-
-Future<void> _tapCombine(WidgetTester tester) async {
-  final combine = find.widgetWithText(CtActionTextButton, 'Combine');
-  await tester.scrollUntilVisible(combine, 120);
-  await tester.pumpAndSettle();
-  await tester.tap(combine);
-  await tester.pumpAndSettle();
-}
-
-void _expectCombineEnabled(WidgetTester tester, {required bool enabled}) {
-  expect(
-    tester
-        .widget<CtActionTextButton>(
-          find.widgetWithText(CtActionTextButton, 'Combine'),
-        )
-        .enabled,
-    enabled,
-  );
-}
-
-(AppEventBus, NavalFleetsUpdatedEvent? Function()) _wireFleetsUpdated() {
-  NavalFleetsUpdatedEvent? updated;
-  final bus = AppEventBus.create();
-  addTearDown(
-    bus.on<NavalFleetsUpdatedEvent>().listen((e) => updated = e).cancel,
-  );
-  return (bus, () => updated);
-}
-
-Fleet _pf(
-  String id,
-  String humanId,
-  String shipId,
-  String typeId, {
-  FleetMission mission = FleetMission.none,
-}) => Fleet(
-  id: id,
-  ownerId: humanId,
-  regionId: 'oldWorld',
-  inPortAtProvinceId: _mergePort,
-  ships: [ShipInstance(id: shipId, typeId: typeId)],
-  mission: mission,
-);
-
-Game _mergePortGame({
-  required String humanId,
-  required String gameId,
-  required String displayName,
-  required List<Fleet> fleets,
-  bool playerHasCapital = true,
-  int nextShipInstanceSeq = 3,
-}) => buildNavalPanelCapitalMergePortFleetsGame(
-  humanId: humanId,
-  gameId: gameId,
-  displayName: displayName,
-  playerHasCapital: playerHasCapital,
-  nextShipInstanceSeq: nextShipInstanceSeq,
-  fleets: fleets,
-);
-
-Game _sameSeaCombineGame({required String humanId}) {
-  const capProvince = 'oldWorld|cap1';
-  return buildNavalPanelOwFleetsGame(
-    gameId: 'g_same_sea_combine',
-    humanId: humanId,
-    displayName: 'Same-sea combine',
-    capitalProvinceId: capProvince,
-    oldWorldProvinces: [
-      Province(
-        id: 'coast',
-        regionId: 'oldWorld',
-        ownerId: humanId,
-        displayName: 'Coast',
-      ),
-      Province(
-        id: 'cap1',
-        regionId: 'oldWorld',
-        ownerId: humanId,
-        displayName: 'Capital',
-      ),
-    ],
-    fleets: [
-      Fleet(
-        id: 'sea_1',
-        ownerId: humanId,
-        regionId: 'oldWorld',
-        seaZoneId: 'zone_alpha',
-        inPortAtProvinceId: null,
-        ships: const [ShipInstance(id: 'ss1', typeId: 'carrack')],
-        mission: FleetMission.patrol,
-      ),
-      Fleet(
-        id: 'sea_2',
-        ownerId: humanId,
-        regionId: 'oldWorld',
-        seaZoneId: 'zone_alpha',
-        inPortAtProvinceId: null,
-        ships: const [ShipInstance(id: 'ss2', typeId: 'fluyte')],
-      ),
-    ],
-    portsByProvinceSeaboard: {
-      'oldWorld|coast|zone_alpha': 'oldWorld|coast|0|0',
-    },
-    tileKeysByProvince: {
-      capProvince: ['oldWorld|cap1|0|0'],
-      'oldWorld|coast': ['oldWorld|coast|0|0'],
-    },
-    nextShipInstanceSeq: 3,
-  );
-}
-
-Future<NavalFleetsUpdatedEvent?> _pumpCheckCombine(
-  WidgetTester tester, {
-  required Game game,
-  required String humanId,
-  required List<String> labels,
-  bool scroll = false,
-  bool? expectCombineEnabled,
-}) async {
-  final (bus, latest) = _wireFleetsUpdated();
-  await pumpNavalPanel(tester, game: game, humanPlayerId: humanId, bus: bus);
-  await _tapFleetCheckboxes(tester, labels, scroll: scroll);
-  if (expectCombineEnabled != null) {
-    _expectCombineEnabled(tester, enabled: expectCombineEnabled);
-  }
-  await _tapCombine(tester);
-  return latest();
-}
 
 void main() {
   suppressLogsForTests();
@@ -226,8 +66,8 @@ void main() {
           bus: bus,
         );
 
-        await _tapFleetCheckboxes(tester, ['Home Fleet', 'Fleet sea_source']);
-        await _tapCombine(tester);
+        await tapNavalFleetCheckboxes(tester, ['Home Fleet', 'Fleet sea_source']);
+        await tapNavalCombine(tester);
 
         final moveOneFluyte = find.byKey(
           CtTransferListKeys.leftMoveOne('fluyte'),
@@ -293,8 +133,8 @@ void main() {
           ),
         );
 
-        await _tapFleetCheckboxes(tester, ['Home Fleet', 'Fleet sea_far']);
-        _expectCombineEnabled(tester, enabled: false);
+        await tapNavalFleetCheckboxes(tester, ['Home Fleet', 'Fleet sea_far']);
+        expectNavalCombineEnabled(tester, enabled: false);
       },
     );
 
@@ -302,9 +142,9 @@ void main() {
       WidgetTester tester,
     ) async {
       const sameSeaId = 'gp_same_sea_combine';
-      final sameSea = await _pumpCheckCombine(
+      final sameSea = await pumpNavalTapCheckCombine(
         tester,
-        game: _sameSeaCombineGame(humanId: sameSeaId),
+        game: buildNavalPanelSameSeaCombineGame(humanId: sameSeaId),
         humanId: sameSeaId,
         labels: const ['Fleet sea_1', 'Fleet sea_2'],
         expectCombineEnabled: true,
@@ -319,21 +159,21 @@ void main() {
       expect(sameSeaSurvivor.mission, FleetMission.none);
 
       const missionId = 'gp_mission_clear';
-      final cleared = await _pumpCheckCombine(
+      final cleared = await pumpNavalTapCheckCombine(
         tester,
-        game: _mergePortGame(
+        game: buildNavalPanelMergePortFleetsGame(
           humanId: missionId,
           gameId: 'g_mission_clear',
           displayName: 'Mission clear tester',
           fleets: [
-            _pf(
+            navalPanelPortFleetAtMergePort(
               'm1',
               missionId,
               'ms1',
               'carrack',
               mission: FleetMission.patrol,
             ),
-            _pf(
+            navalPanelPortFleetAtMergePort(
               'm2',
               missionId,
               'ms2',
@@ -357,17 +197,16 @@ void main() {
       (WidgetTester tester) async {
         const humanId = 'gp_partial_header';
 
-        // No capital => no synthetic Home Fleet row; select-all stays one locality.
-        final partialGame = _mergePortGame(
+        final partialGame = buildNavalPanelMergePortFleetsGame(
           humanId: humanId,
           gameId: 'g_partial_header',
           displayName: 'Partial header tester',
           playerHasCapital: false,
           nextShipInstanceSeq: 4,
           fleets: [
-            _pf('p1', humanId, 'ps1', 'carrack'),
-            _pf('p2', humanId, 'ps2', 'fluyte'),
-            _pf('p3', humanId, 'ps3', 'carrack'),
+            navalPanelPortFleetAtMergePort('p1', humanId, 'ps1', 'carrack'),
+            navalPanelPortFleetAtMergePort('p2', humanId, 'ps2', 'fluyte'),
+            navalPanelPortFleetAtMergePort('p3', humanId, 'ps3', 'carrack'),
           ],
         );
 
@@ -400,7 +239,7 @@ void main() {
           expect(tester.widget<Checkbox>(cb).value, isTrue);
         }
 
-        _expectCombineEnabled(tester, enabled: true);
+        expectNavalCombineEnabled(tester, enabled: true);
       },
     );
 
@@ -408,17 +247,17 @@ void main() {
       'AC: Three-fleet combine survivor is first in panel order regardless of check order',
       (WidgetTester tester) async {
         const humanId = 'gp_reverse_check';
-        final updated = await _pumpCheckCombine(
+        final updated = await pumpNavalTapCheckCombine(
           tester,
-          game: _mergePortGame(
+          game: buildNavalPanelMergePortFleetsGame(
             humanId: humanId,
             gameId: 'g_reverse_check',
             displayName: 'Reverse check tester',
             nextShipInstanceSeq: 4,
             fleets: [
-              _pf('r1', humanId, 'rs1', 'carrack'),
-              _pf('r2', humanId, 'rs2', 'fluyte'),
-              _pf('r3', humanId, 'rs3', 'carrack'),
+              navalPanelPortFleetAtMergePort('r1', humanId, 'rs1', 'carrack'),
+              navalPanelPortFleetAtMergePort('r2', humanId, 'rs2', 'fluyte'),
+              navalPanelPortFleetAtMergePort('r3', humanId, 'rs3', 'carrack'),
             ],
           ),
           humanId: humanId,
@@ -439,15 +278,16 @@ void main() {
       (WidgetTester tester) async {
         const humanId = 'gp_prune_sel';
 
-        Game gameWithKeepDrop(String keep, String drop) => _mergePortGame(
-          humanId: humanId,
-          gameId: 'g_prune_two',
-          displayName: 'Prune tester',
-          fleets: [
-            _pf(keep, humanId, 'ks1', 'carrack'),
-            _pf(drop, humanId, 'ks2', 'fluyte'),
-          ],
-        );
+        Game gameWithKeepDrop(String keep, String drop) =>
+            buildNavalPanelMergePortFleetsGame(
+              humanId: humanId,
+              gameId: 'g_prune_two',
+              displayName: 'Prune tester',
+              fleets: [
+                navalPanelPortFleetAtMergePort(keep, humanId, 'ks1', 'carrack'),
+                navalPanelPortFleetAtMergePort(drop, humanId, 'ks2', 'fluyte'),
+              ],
+            );
 
         final gameTwo = gameWithKeepDrop('stays', 'removed');
         final gameOne = gameTwo.copyWith(
@@ -460,7 +300,7 @@ void main() {
         );
 
         await pumpNavalPanel(tester, game: gameTwo, humanPlayerId: humanId);
-        await _tapFleetCheckboxes(tester, ['Fleet stays', 'Fleet removed']);
+        await tapNavalFleetCheckboxes(tester, ['Fleet stays', 'Fleet removed']);
 
         await pumpNavalPanel(tester, game: gameOne, humanPlayerId: humanId);
 
@@ -474,7 +314,7 @@ void main() {
           findsNothing,
         );
         expect(tester.widget<Checkbox>(staysCb).value, isTrue);
-        _expectCombineEnabled(tester, enabled: false);
+        expectNavalCombineEnabled(tester, enabled: false);
       },
     );
 
@@ -482,17 +322,17 @@ void main() {
       'AC: Collapsed rows keep inline Split action while checkbox selection works',
       (WidgetTester tester) async {
         const humanId = 'gp_collapsed_cb';
-        final collapsedGame = _mergePortGame(
+        final collapsedGame = buildNavalPanelMergePortFleetsGame(
           humanId: humanId,
           gameId: 'g_collapsed_cb',
           displayName: 'Collapsed cb tester',
           fleets: [
-            _pf('col_a', humanId, 'cs1', 'carrack'),
-            _pf('col_b', humanId, 'cs2', 'fluyte'),
+            navalPanelPortFleetAtMergePort('col_a', humanId, 'cs1', 'carrack'),
+            navalPanelPortFleetAtMergePort('col_b', humanId, 'cs2', 'fluyte'),
           ],
         );
 
-        final (bus, latest) = _wireFleetsUpdated();
+        final (bus, latest) = wireNavalFleetsUpdatedCapture();
         await pumpNavalPanel(
           tester,
           game: collapsedGame,
@@ -509,12 +349,12 @@ void main() {
           );
         }
 
-        await _tapFleetCheckboxes(tester, ['Fleet col_a', 'Fleet col_b']);
+        await tapNavalFleetCheckboxes(tester, ['Fleet col_a', 'Fleet col_b']);
         expect(
           find.descendant(of: tileA, matching: find.byTooltip('Split')),
           findsOne,
         );
-        await _tapCombine(tester);
+        await tapNavalCombine(tester);
 
         final updated = latest();
         expect(updated, isNotNull);
